RAG Project/RAG Recipe/requirements.txt
# RAG Recipe - FastAPI UI for RAG Pipelines
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.2
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# RAG & LLM
openai>=1.12.0
requests>=2.31.0
sentence-transformers>=2.2.2
chromadb>=0.4.22
tiktoken>=0.5.2

# Document processing
pypdf2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.1.0
python-magic-bin>=0.4.14; sys_platform == 'win32'

# Utilities
aiofiles>=23.2.1
